Network Error fix
There was an issue when attempting to make a request to the api regarding CLEARTEXT error. Since the API it's not under https yet, this error was showing up and making it impossible to send the resquest. A xml network-config file was added with the domain in order to allow cleartext by now.